Generator Diety i test dodawania cwiczen do dnia

+++ src/main/java/workout/workoutapp/database/entities/Diet.java
@@ -15,16 +15,16 @@ public class Diet {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long diet_id;
-    private Long kcal;
-    private Long protein;
-    private Long fat;
-    private Long carbohydrates;
+    private float kcal;
+    private float protein;
+    private float fat;
+    private float carbohydrates;
 
     @ManyToOne
     @JoinColumn(name="user_id")
     private User user;
 
-    public Diet(Long kcal, Long protein, Long fat, Long carbohydrates, User user) {
+    public Diet(float kcal, float protein, float fat, float carbohydrates, User user) {
         this.kcal = kcal;
         this.protein = protein;
         this.fat = fat;
